update employee role function

+++ data.sql
@@ -16,7 +16,7 @@ VALUES ('Sales'),
 CREATE TABLE roles (
     role_id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     role VARCHAR(50) NOT NULL,
-    salary INT NOT NULL,
+    salary DECIMAL NOT NULL,
     department_id INT NOT NULL,
     FOREIGN KEY (department_id) REFERENCES departments(department_id)
 );
